plot/requirements.txt
matplotlib
numpy
pyqtgraph
pyserial
PySide6
termcolor
